Refactored AllViewTest helpers
